refactor: MVC 구조로 수정
